refactor: 달력 모드에서만 사용하는 element와 하루 모드에서도 사용할 element를 분리
